+++ NEWS.md
@@ -5,8 +5,8 @@
 * `iHndlFormula()`: Modified.  COrrected the positioning of the explanatory variables when the model has a response variable.
 * `iMakeBaseResidPlot()`:  Added as an internal function to `residPlot()` to simplify some coding.
 * `iMakeColor()`: Modified.  More intelligently handles values that are greater than 1 (converts them to decimals by inverting.)
-* `lwPredsComp()`: Modified.  Added use of internal `iHndlFormula()` and moved two internal functions outside the main function.  Changed default for intervals from `both` to `confidence` and changed so that if only the confidence or prediction intervals are plotted they will be black with `lwd=` width (if both are plotted the CI is now black and the PI is now blue).  Added a `show.preds` argument.  Changed `connect.means=` to `connect.preds=`.  Changed default `lwd=` value and how it is used for CIs, PIs, and the connection lines.  Added `col.connect=` argument.  Removed `mar` and `mgp` from `par()` call (left `mfrow`).  Added more examples. 
-* `residPlot()`: Modified.  Corrected a bug around the use of `thigmophobe()` in `iAddOutlierTest()`.
+* `lwPredsComp()`: Modified.  Changed `mdl=` to `object=`.  Added use of internal `iHndlFormula()` and moved two internal functions outside the main function.  Changed default for intervals from `both` to `confidence` and changed so that if only the confidence or prediction intervals are plotted they will be black with `lwd=` width (if both are plotted the CI is now black and the PI is now blue).  Added a `show.preds` argument.  Changed `connect.means=` to `connect.preds=`.  Changed default `lwd=` value and how it is used for CIs, PIs, and the connection lines.  Added `col.connect=` argument.  Removed `mar` and `mgp` from `par()` call (left `mfrow`).  Added more examples.  Added tests for error messages.
+* `residPlot()`: Modified.  Added `inclHist=` argument.  Corrected a bug around the use of `thigmophobe()` in `iAddOutlierTest()`.  Changed default for `student=` to `FALSE`.  Modified and added more examples.
 * `SMBassWB`: Modified.  Added a seealso.